#135: Splitting out the lookup of passwords in config to support same value in config, and vault config path resolution.
------
In order to activate the vault resolution, you are only required to supply the vault configuration in the yml file, otherwise just the default password resolution(which just returns the same value) will be used. It has been done like this to enable different resolution strategies if we so desire in the future.